Estimate locations

Calculate point estimates based on antenna bearing and signal strength.

locate(
  data,
  ID = "tagDeployID",
  ts = "ts",
  sig = "sig",
  aLon = "recvDeployLon",
  aLat = "recvDeployLat",
  aType = "antType",
  aBearing = "antBearing",
  det_range = 12,
  dtime = 2
)

Arguments

data

A data.frame containing the telemetry data.

ID

Unique identifier for individuals or tag deployments.

ts

Timestamp column.

sig

Signal strength column.

aLon

Antenna longitude column.

aLat

Antenna latitude column.

aType

Antenna type column.

aBearing

Antenna bearing column.

det_range

Assumed maximum detection range of antennas in kilometres. Can be a single value or a named list of values for different antenna types.

dtime

Time interval in minutes for which point estimates are to be calculated.

Value

Returns a data.frame containing estimated coordinates and measurement errors for each time interval together with the proportions of time intervals w.

Details

This function performs the following steps as described in Baldwin et al. 2018:

Estimate locations for each detection: half of the maximum detection range det_range along the directional beam.

Derive oscillating measurement error arising from antenna geometry and orientation.

Calculate weighted means (by signal strength) for each time interval dtime.

Examples

if (FALSE) {
data(motusData)
locate(motusData)
locate(motusData, dtime = 1, det_range = 10)
locate(motusData, det_range = list("yagi-5"=10, "yagi-6"=12))
}