Model data

Model flight path from point estimates using a DCRW model.

track(data, method = "mcmc", ...)

Arguments

data

A data.frame containing the point estimate data.

method

The estimation method to use. Either "mcmc" or "optim"; defaults to "mcmc".

...

Additional arguments passed to cmdstanr::sample() or cmdstanr::optimize(), respectively.

Value

Returns a data.frame containing estimates for longitude, latitude, distance and speed per time interval or a stantrackr object including the full posterior distributions.

Details

This function calls Stan via cmdstanr and uses a difference correlated random walk model (DCRW) to estimate individual flight paths. The model is described in more detail in Jonsen et al. 2005 and Baldwin et al. 2018. To learn more about state-space models in animal movement in general, Auger-Méthé et al. 2021 is a good starting point. Available estimation methods are Stan's main Markov chain Monte Carlo algorithm and Stan's optimizer to obtain a (penalized) maximum likelihood estimate or a maximum a posteriori estimate (if jacobian=TRUE). See the CmdStan User's Guide for more details.

See also

Examples

if (FALSE) {
# Load data
data(motusData)

# Estimate locations
loc <- locate(motusData, dtime = 2)

# Model flight paths
track(loc, parallel_chains = 4)
track(loc, "optim", refresh = 1e3)
}