Model data

Model flight path from point estimates using a DCRW model.

track(data = NULL, ci = "HDI", prob = 0.9, output_dir = getwd(), ...)

Arguments

data

A data.frame containing the point estimate data.

ci

The method used to calculate the credible intervals. Available options are 'HDI' for the highest posterior density interval and 'ETI' for the equal-tailed interval; defaults to 'HDI'.

prob

The probability mass of the credible interval; defaults to 0.9.

output_dir

The directory to save the CmdStanMCMC objects to; defaults to the current working directory. If NULL, the objects are not saved.

...

Additional arguments passed to cmdstanr::sample().

Value

Returns a data.frame containing the model summaries.

Details

This function is a wrapper around cmdstanr::sample() and uses a difference correlated random walk model (DCRW) to estimate individual flight paths. The model is described in more detail in Jonsen et al. 2005 and Baldwin et al. 2018. To learn more about state-space models in animal movement in general, Auger-Méthé et al. 2021 is a good starting point.